# src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: Ohgnoy API
  version: 1.0.0
  description: Ohgnoy 백엔드 API 문서

servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버

paths:
  /user/register:
    post:
      summary: 사용자 회원가입
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

  /user/login:
    post:
      summary: 사용자 로그인
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginRequest"
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

  /post/list:
    get:
      summary: 게시글 목록 조회
      tags:
        - Post
      responses:
        "200":
          description: 게시글 목록 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /post/create:
    post:
      summary: 게시글 생성
      tags:
        - Post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Post"
      responses:
        "200":
          description: 게시글 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        password:
          type: string
        email:
          type: string

    UserLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string

    ApiResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/UserResponse"

    Post:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        author:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time